PickFur: single-precision rayStartInit x/y
Compute ray-start NDC x/y in single precision (fsubs/fdivs) instead
of nested double casts (fsub/fdiv+frsp), matching target.
76.55 -> 76.60%, PickFur 55.05 -> 55.36%.

+++ src/chara_fur.cpp
@@ -1438,10 +1438,8 @@ int CChara::CModel::PickFur(
 						CVector rayStart;
 						CVector rayEnd;
 						CVector rayStartInit(
-						    static_cast<float>(static_cast<double>(static_cast<float>(cursorXd -
-						                                                              static_cast<double>(kCharaFurScreenCenterX))) /
-						                       static_cast<double>(kCharaFurScreenCenterX)),
-						    static_cast<float>(negCursorY / static_cast<double>(kCharaFurScreenCenterY)), kCharaFurDepthZero);
+						    (static_cast<float>(cursorXd) - kCharaFurScreenCenterX) / kCharaFurScreenCenterX,
+						    static_cast<float>(negCursorY) / kCharaFurScreenCenterY, kCharaFurDepthZero);
 						rayStart.x = rayStartInit.x;
 						rayStart.y = rayStartInit.y;
 						rayStart.z = rayStartInit.z;
